requirements.txt
streamlit==1.31.0
pandas==2.2.0
praw==7.7.1
feedparser==6.0.11
plotly==5.18.0
python-dotenv==1.0.0
